Issue 491, use long instead of int for the timeout related APIs. Merged Nandor Kracser <bonifaido@gmail.com> changes plus some minor changes.

+++ src/org/mockito/internal/verification/VerificationOverTimeImpl.java
@@ -16,8 +16,8 @@ import org.mockito.verification.VerificationMode;
  */
 public class VerificationOverTimeImpl implements VerificationMode {
 
-    private final int pollingPeriodMillis;
-    private final int durationMillis;
+    private final long pollingPeriodMillis;
+    private final long durationMillis;
     private final VerificationMode delegate;
     private final boolean returnOnSuccess;
     
@@ -32,7 +32,7 @@ public class VerificationOverTimeImpl implements VerificationMode {
      *                        the delegate is satisfied and the full duration has passed (as in
      *                        {@link org.mockito.verification.VerificationAfterDelay}).
      */
-    public VerificationOverTimeImpl(int pollingPeriodMillis, int durationMillis, VerificationMode delegate, boolean returnOnSuccess) {
+    public VerificationOverTimeImpl(long pollingPeriodMillis, long durationMillis, VerificationMode delegate, boolean returnOnSuccess) {
         this.pollingPeriodMillis = pollingPeriodMillis;
         this.durationMillis = durationMillis;
         this.delegate = delegate;
@@ -93,7 +93,7 @@ public class VerificationOverTimeImpl implements VerificationMode {
         return !(verificationMode instanceof AtMost || verificationMode instanceof NoMoreInteractions);
     }
 
-    private void sleep(int sleep) {
+    private void sleep(long sleep) {
         try {
             Thread.sleep(sleep);
         } catch (InterruptedException ie) {
@@ -101,11 +101,11 @@ public class VerificationOverTimeImpl implements VerificationMode {
         }
     }
     
-    public int getPollingPeriod() {
+    public long getPollingPeriod() {
         return pollingPeriodMillis;
     }
     
-    public int getDuration() {
+    public long getDuration() {
         return durationMillis;
     }
